Add support for SSE KMS encryption is AWS S3 output
When SSE KMS encryption is enabled for an S3 bucket the id of the
KMS key to use to encrypt the data can be provided at the time of
uploading an object.

Additionally some environments require a KMS key id to be provided
if the encrytion of objects in the given bucket is mandatory.

Added support for configuring an SSE KMS key id as part of the request
to publish an object when the AWS S3 output is enabled.

NOTE: We can't contribute this patch to upstream because the current
supported version of falcosidekick (2.32.0) migrated from aws-sdk-go
to aws-sdk-go-2 and we can't adopt 2.32.0 because of an existing bug
(https://github.com/falcosecurity/falcosidekick/issues/1191)